rename definitions.py to connectors_definitions.py
